Move device removal dialog to dialog package
